+++ libs/cardano-ledger-pretty/src/Cardano/Ledger/Pretty/Conway.hs
@@ -18,7 +18,7 @@ module Cardano.Ledger.Pretty.Conway (
 ) where
 
 import Cardano.Ledger.Alonzo.Scripts (CostModels, ExUnits, Prices)
-import Cardano.Ledger.BaseTypes (EpochNo, NonNegativeInterval, UnitInterval)
+import Cardano.Ledger.BaseTypes (EpochInterval (..), EpochNo, NonNegativeInterval, UnitInterval)
 import Cardano.Ledger.Coin (Coin)
 import Cardano.Ledger.Compactible (fromCompact)
 import Cardano.Ledger.Conway (ConwayEra)
@@ -305,6 +305,7 @@ ppConwayPParams ::
   , PrettyA (HKD f Natural)
   , PrettyA (HKD f ExUnits)
   , PrettyA (HKD f EpochNo)
+  , PrettyA (HKD f EpochInterval)
   , PrettyA (HKD f DRepVotingThresholds)
   , PrettyA (HKD f CostModels)
   , PrettyA (HKD f CoinPerByte)