+++ eras/babbage/impl/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Version history for `cardano-ledger-babbage`
 
-## 1.5.1.1
+## 1.6.0.0
 
-*
+* Change the type of `ppEMaxL`
 
 ## 1.5.1.0